Equipment: Armor

+++ src/main/java/com/example/charactergenerator/repository/ArmorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArmorRepository extends CrudRepository<Armor,Byte> {
+public interface ArmorRepository extends CrudRepository<Armor,Long> {
 }
